+++ CHANGES.rst
@@ -5,6 +5,8 @@ Change log
 ----------------
 
 
+- Adds invalidate method to MemCacheMapping and test
+
 4.0 (2023-02-02)
 ----------------